Refactor PostCategory to use categoryName parameter
Updated the PostCategory method in CategoryController to accept a string parameter `categoryName` instead of a `Category` object. The null check now verifies `categoryName`, and a new `Category` object is created from it before being added to the database. The response message has been adjusted to reflect the added category's name.